It works
data acurately displays on the globe and updates when a year is
selected. I am going to implement an update so that the globe
repopulates it data when a new country is selected also. If data points
are able to be selected a future update will allow the user to select a
point and see more information about it